% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_utils.R
\name{extract_audio_chunk}
\alias{extract_audio_chunk}
\title{Extracts a section (chunk) of an audio file}
\usage{
extract_audio_chunk(path, start, duration, quiet = TRUE)
}
\arguments{
\item{path}{the path to the file}

\item{start}{the staring point, in seconds}

\item{duration}{the length of the chunk, in seconds}

\item{quiet}{logical, whether to hide the verbose output}
}
\value{
a \link[tuneR:Wave]{tuneR::Wave} object corresponding to the parsed chunk
}
\description{
Read an arbitrary section of an mp3 file as a Wave object
}
