% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_utils.R
\name{file_duration}
\alias{file_duration}
\title{Retrieve total audio file duration}
\usage{
file_duration(path)
}
\arguments{
\item{path}{the path to the file}
}
\value{
a number of seconds
}
\description{
Uses ffprobe to estimate the duration of an audio resource
}
