% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_audio.R
\name{map_mp3_chunks}
\alias{map_mp3_chunks}
\title{Apply an arbitrary function on successive chunks of an audio file}
\usage{
map_mp3_chunks(
  path,
  FUN,
  chunk_duration,
  show_progress = FALSE,
  cache_dir = NULL,
  ffmpeg_quiet = TRUE,
  ...
)
}
\arguments{
\item{path}{the path to the file}

\item{FUN}{a function to be run on each chunk, see detail}

\item{chunk_duration}{the length of the chunk, in seconds}

\item{show_progress}{logical, whether to display progress}

\item{cache_dir}{an optional directory, where the chunk is indefinitely stored}

\item{ffmpeg_quiet}{logical, whether to hide the verbose output from ffmpeg}

\item{...}{arguments to be passed to \code{FUN}}
}
\value{
a data.table where each row is a chunk. columns are the start time, in second, and the variables defined by \code{FUN}
}
\description{
Map a user-defined function to a series of consecutive windows of arbitrary duration.
}
\details{
The function must use a \link[tuneR:Wave]{tuneR::Wave} as an input and output a named list. Each element of the list will be parsed as a new column in the output
}
