% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp3_utils.R
\name{mp3_extract_chunk}
\alias{mp3_extract_chunk}
\title{Extracts a section (chunk) of an mp3 file}
\usage{
mp3_extract_chunk(path, start, duration, cache_dir = NULL, quiet = TRUE)
}
\arguments{
\item{path}{the path to the file}

\item{start}{the staring point, in seconds}

\item{duration}{the length of the chunk, in seconds}

\item{cache_dir}{an optional directory, where the chunk is indefinitely stored (see details)}

\item{quiet}{logical, whether to hide the verbose output}
}
\value{
a \link[tuneR:Wave]{tuneR::Wave} object corresponding to the parsed chunk
}
\description{
Read an arbitrary section of an mp3 file as a Wave object
}
\details{
since seeking and clipping mp3 is computationally consuming using cache (\code{cache_dir}) \
stores the resulting mp3 section. Caching hashes the path of the file as well as the chunk start and duration.
In other words, if the path is changed, or a new duration/starting point is requested, a new mp3 will be generated.
One can eventually delete the cache directory to free space.
}
