% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{tempaural_example}
\alias{tempaural_example}
\title{Get path to tempaural example}
\usage{
tempaural_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
\code{tempaural} comes with a sample audio files in its \code{inst/extdata}
directory. \code{tempaural} allow  make them easy to access.
}
\examples{
# list all files
scopr_example()
}
